Fix build errors
fmt with C++20 enabled now has compile time checks (if possible) of the format string against parameter types. This can detect formatting problems.

Not sure if this would actually have resulted in problem when formatting the enums, but it can't hurt to explicitly cast them to a integer type as that is the desired way to format them in those places that I adjusted.

+++ engine/sim/expressions.cpp
@@ -1147,7 +1147,7 @@ void print_tokens( util::span<const expr_token_t> tokens, sim_t* sim )
 
   std::vector<std::string> strings;
   for ( const expr_token_t& token : tokens )
-    strings.push_back( fmt::format("{:2d} '{}'", token.type, token.label ) );
+    strings.push_back( fmt::format("{:2d} '{}'", static_cast<int>( token.type), token.label ) );
 
   sim->out_debug.print( "{}", fmt::join( strings, " | " ) );
 }
